requirements.txt
flask
flask-sqlalchemy
flask_wtf
flask-Login
flask-bcrypt
flask-migrate
itsdangerous
pandas
matplotlib
ffmpeg-python